refactor: Refactors error handling and simplifies imports
