+++ api/core/app/apps/exc.py
@@ -1,0 +1,2 @@
+class GenerateTaskStoppedError(Exception):
+    pass
