+++ api/core/workflow/nodes/http_request/node.py
@@ -82,7 +82,7 @@ class HttpRequestNode(BaseNode):
 
             response = http_executor.invoke()
             files = self.extract_files(url=http_executor.url, response=response)
-            if not response.response.is_success and (self.should_continue_on_error or self.should_retry):
+            if not response.response.is_success and (self.continue_on_error or self.retry):
                 return NodeRunResult(
                     status=WorkflowNodeExecutionStatus.FAILED,
                     outputs={
@@ -221,3 +221,11 @@ class HttpRequestNode(BaseNode):
         files.append(file)
 
         return ArrayFileSegment(value=files)
+
+    @property
+    def continue_on_error(self) -> bool:
+        return self.node_data.error_strategy is not None
+
+    @property
+    def retry(self) -> bool:
+        return self.node_data.retry_config.retry_enabled
